+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## [10.3.0] - 2020-01-07
+
+## Added
+- Editable Table component
+- Keyboard input to Linn Week Picker
+
+## Changed
+- Linn Week Picker component label prop is no longer required
+- Dropdown label prop is no longer required
+
 ## [10.2.5] - 2020-01-06
 
 ## Changed